Review changes to interrupted sync, base, and start date tests
